CJ Orders
This project retrieves Return/Refund information for our E-commerce company. The program is written in Python and it retrieves this information using 2 GraphQL API queries. It uses the API connections to connect to 2 different websites. It retrieves a JSON file containing product information from each website. I can extract the information I need from the JSON files using a Python for Loop, and then save what I need into a list. Using the Pandas Library within Python I take the list and convert it into a DataFrame and export it as a well-organized and readable excel file for the user to examine.