Remove Order Detail
This project checks to see if a product has been discontinued on Amazon Seller, it then updates the company’s database to show the status that is on Amazon.  The program is written in Python. It connects to Amazon using a MWS API connection. It then retrieves a .TXT file containing the SKUs of products that have been discontinued. The program reads the .TXT file, and extracts the SKUs using a For Loop. It then saves the SKUs to a list. Using another For Loop I can loop through that list and run a SQL query for each SKU. The query is a UPDATE query that connects to the Inventory table in the company’s Access Database and updates the status for each SKU to match the status that is in Amazon. After the program finishes, the statuses in the database are up to date and match the statuses on Amazon Seller.