Revenue Checker
I wanted to create a quick way for someone to check the company’s Total Daily Revenue. The program is written in Python and uses SQL to connect to the company’s Database.  The program uses a GUI interface that requests a Date from the user. Once the user inputs a Date, the program runs a SQL query to connect to the Orders table in the Company’s Access Database. The SQL query retrieves columns containing sales information for that specific Date. Using the Pandas library within Python, the program places the information within a DataFrame, and then calculates the total revenue using a custom formula.  The program then displays the Total Daily Revenue in a GUI interface for the user to view. I have also created 2 different variations of this program, one that displays the Total Monthly Revenue, and another that allows you to look for the revenue of a specific order by inputting the order’s OrderNumber.